German-Traffic-Sign-Classification-App/requirements.txt
streamlit==1.14.0
Pillow==10.3.0
numpy==1.26.4
tensorflow==2.16.1
